requirements.txt
certifi==2019.11.28
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
kiwisolver==1.1.0
matplotlib==3.2.0
networkx==2.4
numpy==1.18.1
Pillow==7.0.0
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
six==1.14.0
tensorboardX==2.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
yacs==0.1.6
